+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ## Helpers
 
 * A new humanizeFilesize helper for converting bytes to a human readable representation.
+* `roundForAbbreviation` is added to handle the number element of `abbreviateNumber` as well as forcing any abbreviated number to one decimal place
 
 # 0.28.2